Fix moving into negative coordinates

+++ src/main/java/me/zero/simulatimod/mixin/entity/PlayerEntityMixin.java
@@ -9,11 +9,11 @@ import org.spongepowered.asm.mixin.injection.ModifyConstant;
 public class PlayerEntityMixin {
     @ModifyConstant(method = "tick", constant = @Constant(doubleValue = -2.9999999E7))
     public double simulatiMod$modifyConstantTickNegative(double original) {
-        return Double.MIN_VALUE;
+        return Long.MIN_VALUE;
     }
 
     @ModifyConstant(method = "tick", constant = @Constant(doubleValue = 2.9999999E7))
     public double simulatiMod$modifyConstantTickPositive(double original) {
-        return Double.MAX_VALUE;
+        return Long.MAX_VALUE;
     }
 }
